backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx>=0.27.0
openai>=1.52.0
pinecone-client>=3.0.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
langgraph>=0.2.0
